move id & nonce from u64 to u16

+++ include/core/protocols/eth/peer.h
@@ -30,9 +30,9 @@ private:
   KnownCache *knownTxs;
 
 public:
-  uint64_t id;
+  uint16_t id;
 
-  Peer(uint64_t id) {
+  Peer(uint16_t id) {
     this->id = id;
     this->knownBlocks = new KnownCache(maxKnownBlocks);
     this->knownTxs = new KnownCache(maxKnownTxs);
