+++ include/emu/config.h
@@ -18,7 +18,7 @@ struct Config {
   uint64_t period;
   uint64_t minDelay, maxDelay;
   uint64_t minTxInterval, maxTxInterval;
-  std::unordered_map<uint64_t, std::unique_ptr<EmuNode>> nodes;
+  std::unordered_map<uint16_t, std::unique_ptr<EmuNode>> nodes;
 };
 
 extern Config global;
