+++ include/event/body_event.h
@@ -4,12 +4,12 @@
 
 class BodyEvent : public Event {
 public:
-  const uint64_t from;
-  const uint64_t to;
+  const uint16_t from;
+  const uint16_t to;
   const uint64_t blockHash;
   const std::vector<std::shared_ptr<Transaction>> txs;
 
-  BodyEvent(uint64_t timestamp, uint64_t from, uint64_t to, uint64_t blockHash,
+  BodyEvent(uint64_t timestamp, uint16_t from, uint16_t to, uint64_t blockHash,
             const std::vector<std::shared_ptr<Transaction>> &txs);
 
   void process(std::priority_queue<Event *, std::vector<Event *>, CompareEvent> &queue,
