+++ include/node/node.h
@@ -11,23 +11,23 @@
 #include "core/types/transaction.h"
 
 struct Node {
-  uint64_t id;
-  uint64_t nonce;
+  uint16_t id;
+  uint16_t nonce;
   uint64_t current;
   std::unordered_map<uint64_t, std::shared_ptr<Transaction>> txPool;
   std::unordered_map<uint64_t, std::shared_ptr<Block>> blocksByNumber;
   std::unordered_map<uint64_t, std::shared_ptr<Block>> blocksByHash;
-  std::vector<uint64_t> peerList;
-  std::unordered_map<uint64_t, Peer *> peerMap;
+  std::vector<uint16_t> peerList;
+  std::unordered_map<uint16_t, Peer *> peerMap;
 
   uint64_t nextFetchHeaderTime;
   uint64_t nextFetchBodyTime;
 
-  std::unordered_map<uint64_t, std::vector<uint64_t>> fetchingHeaders;
+  std::unordered_map<uint64_t, std::vector<uint16_t>> fetchingHeaders;
   std::unordered_map<uint64_t, std::shared_ptr<Header>> fetchedHeaders;
-  std::unordered_map<uint64_t, std::vector<uint64_t>> fetchingBodies;
+  std::unordered_map<uint64_t, std::vector<uint16_t>> fetchingBodies;
 
-  Node(uint64_t id) {
+  Node(uint16_t id) {
     this->id = id;
     this->nonce = 0;
     this->current = 0;
@@ -40,7 +40,7 @@ struct Node {
       delete peer;
   }
 
-  uint64_t nextNonce() { return nonce++; }
+  uint16_t nextNonce() { return nonce++; }
   void addPeer(const std::unique_ptr<Node> &node) {
     peerList.push_back(node->id);
     peerMap[node->id] = new Peer(node->id);
