+++ src/event/header_event.cpp
@@ -2,7 +2,7 @@
 #include "emu/config.h"
 #include "event/block_hash_event.h"
 
-HeaderEvent::HeaderEvent(uint64_t timestamp, uint64_t from, uint64_t to, const std::shared_ptr<Header> &header)
+HeaderEvent::HeaderEvent(uint64_t timestamp, uint16_t from, uint16_t to, const std::shared_ptr<Header> &header)
     : Event(timestamp), from(from), to(to), header(header) {}
 
 void HeaderEvent::process(std::priority_queue<Event *, std::vector<Event *>, CompareEvent> &queue,
