+++ src/event/header_req_event.cpp
@@ -2,7 +2,7 @@
 #include "emu/config.h"
 #include "event/header_event.h"
 
-HeaderReqEvent::HeaderReqEvent(uint64_t timestamp, uint64_t from, uint64_t to, uint64_t blockHash)
+HeaderReqEvent::HeaderReqEvent(uint64_t timestamp, uint16_t from, uint16_t to, uint64_t blockHash)
     : Event(timestamp), from(from), to(to), blockHash(blockHash) {}
 
 void HeaderReqEvent::process(std::priority_queue<Event *, std::vector<Event *>, CompareEvent> &queue,
